+++ core/src/main/java/tech/pegasys/web3signer/core/FilecoinRunner.java
@@ -35,6 +35,7 @@ import tech.pegasys.web3signer.signing.config.metadata.AbstractArtifactSignerFac
 import tech.pegasys.web3signer.signing.config.metadata.BlsArtifactSignerFactory;
 import tech.pegasys.web3signer.signing.config.metadata.Secp256k1ArtifactSignerFactory;
 import tech.pegasys.web3signer.signing.config.metadata.interlock.InterlockKeyProvider;
+import tech.pegasys.web3signer.signing.config.metadata.parser.YamlMapperFactory;
 import tech.pegasys.web3signer.signing.config.metadata.parser.YamlSignerParser;
 import tech.pegasys.web3signer.signing.config.metadata.yubihsm.YubiHsmOpaqueDataProvider;
 import tech.pegasys.web3signer.signing.filecoin.FilecoinNetwork;
@@ -150,7 +151,8 @@ public class FilecoinRunner extends Runner {
                     config.getKeyConfigPath(),
                     "yaml",
                     new YamlSignerParser(
-                        List.of(blsArtifactSignerFactory, secpArtifactSignerFactory)));
+                        List.of(blsArtifactSignerFactory, secpArtifactSignerFactory),
+                        YamlMapperFactory.createYamlMapper(config.getKeyStoreConfigFileMaxSize())));
           }
         });
   }