+++ core/src/main/java/tech/pegasys/web3signer/core/config/Config.java
@@ -37,6 +37,8 @@ public interface Config {
 
   Path getKeyConfigPath();
 
+  int getKeyStoreConfigFileMaxSize();
+
   Boolean isMetricsEnabled();
 
   Integer getMetricsPort();