+++ signing/src/test/java/tech/pegasys/web3signer/signing/KeystoreFileManagerTest.java
@@ -13,23 +13,25 @@
 package tech.pegasys.web3signer.signing;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static tech.pegasys.web3signer.signing.config.metadata.parser.YamlSignerParser.YAML_MAPPER;
 
 import tech.pegasys.web3signer.signing.config.metadata.FileKeyStoreMetadata;
 import tech.pegasys.web3signer.signing.config.metadata.SigningMetadata;
+import tech.pegasys.web3signer.signing.config.metadata.parser.YamlMapperFactory;
 
 import java.nio.file.Path;
 import java.util.Map;
 
 import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.dataformat.yaml.YAMLMapper;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 class KeystoreFileManagerTest {
+  private static YAMLMapper YAML_MAPPER = YamlMapperFactory.createYamlMapper();
 
   @Test
   void configurationFilesAreCreated(@TempDir final Path parentDir) throws Exception {
-    new KeystoreFileManager(parentDir)
+    new KeystoreFileManager(parentDir, YAML_MAPPER)
         .createKeystoreFiles("filename", "{\"test\":true}", "password");
 
     final Path metadataYamlFile = parentDir.resolve("filename.yaml");
@@ -43,7 +45,7 @@ class KeystoreFileManagerTest {
 
   @Test
   void yamlContentIsValidFileKeyStoreMetadata(@TempDir final Path parentDir) throws Exception {
-    new KeystoreFileManager(parentDir)
+    new KeystoreFileManager(parentDir, YAML_MAPPER)
         .createKeystoreFiles("filename", "{\"test\":true}", "password");
 
     final Path metadataYamlFile = parentDir.resolve("filename.yaml");
@@ -57,7 +59,7 @@ class KeystoreFileManagerTest {
 
   @Test
   void yamlContentIsNotConverted(@TempDir final Path parentDir) throws Exception {
-    new KeystoreFileManager(parentDir)
+    new KeystoreFileManager(parentDir, YAML_MAPPER)
         .createKeystoreFiles("filename", "{\"test\":true}", "password");
 
     final Path metadataYamlFile = parentDir.resolve("filename.yaml");
@@ -77,7 +79,7 @@ class KeystoreFileManagerTest {
 
   @Test
   void passwordContentsAreWritten(@TempDir final Path parentDir) throws Exception {
-    new KeystoreFileManager(parentDir)
+    new KeystoreFileManager(parentDir, YAML_MAPPER)
         .createKeystoreFiles("filename", "{\"test\":true}", "password");
 
     final Path keystorePasswordFile = parentDir.resolve("filename.password");
@@ -87,7 +89,7 @@ class KeystoreFileManagerTest {
 
   @Test
   void jsonDataIsWritten(@TempDir final Path parentDir) throws Exception {
-    new KeystoreFileManager(parentDir)
+    new KeystoreFileManager(parentDir, YAML_MAPPER)
         .createKeystoreFiles("filename", "{\"test\":true}", "password");
 
     final Path keystoreJsonFile = parentDir.resolve("filename.json");