+++ signing/src/test/java/tech/pegasys/web3signer/signing/config/metadata/InterlockSigningMetadataFileParsingTest.java
@@ -14,9 +14,9 @@ package tech.pegasys.web3signer.signing.config.metadata;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
-import static tech.pegasys.web3signer.signing.config.metadata.parser.YamlSignerParser.YAML_MAPPER;
 
 import tech.pegasys.web3signer.signing.KeyType;
+import tech.pegasys.web3signer.signing.config.metadata.parser.YamlMapperFactory;
 
 import java.io.IOException;
 import java.net.URI;
@@ -24,10 +24,12 @@ import java.net.URL;
 import java.nio.file.Path;
 
 import com.fasterxml.jackson.databind.JsonMappingException;
+import com.fasterxml.jackson.dataformat.yaml.YAMLMapper;
 import com.google.common.io.Resources;
 import org.junit.jupiter.api.Test;
 
 class InterlockSigningMetadataFileParsingTest {
+  private static final YAMLMapper YAML_MAPPER = YamlMapperFactory.createYamlMapper();
 
   @Test
   void yamlFileIsSuccessfullyParsed() throws IOException {